EREGCSC-2038 Put sections in statute table in alphanumeric order (#886)
* Order by U.S.C. update tests

* Add natural sort to usc field

* Update fixtures

* Linting
