+++ solution/backend/common/fields.py
@@ -2,6 +2,7 @@
 
 
 import datetime
+import re
 
 from django.core.exceptions import ValidationError
 from django.core.validators import RegexValidator
@@ -37,3 +38,32 @@ class VariableDateField(models.CharField):
             ],
         }}
         return super().__init__(*args, **kwargs)
+
+
+class NaturalSortField(models.CharField):
+    def __init__(self, for_field, **kwargs):
+        self.for_field = for_field
+        kwargs.setdefault('db_index', True)
+        kwargs.setdefault('editable', False)
+        kwargs.setdefault('max_length', 255)
+        super(NaturalSortField, self).__init__(**kwargs)
+        self.max_length = kwargs['max_length']
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(NaturalSortField, self).deconstruct()
+        args.append(self.for_field)
+        return name, path, args, kwargs
+
+    def pre_save(self, model_instance, add):
+        return self.naturalize(getattr(model_instance, self.for_field))
+
+    def naturalize(self, string):
+        def naturalize_int_match(match):
+            return '%08d' % (int(match.group(0)),)
+        if string:
+            string = string.lower()
+            string = string.strip()
+            string = re.sub(r'\d+', naturalize_int_match, string)
+            string = string[:self.max_length]
+
+        return string
