+++ solution/backend/regulations/tests/fixtures/statute_link_api_test.json
@@ -30,6 +30,16 @@
         "source_url": "https://test.gov/test3.xml"
     },
     {
+        "section": "7890-2",
+        "title": 42,
+        "usc": "77777-1A",
+        "act": "Social Security Act",
+        "name": "Something else (2).",
+        "statute_title": 3,
+        "statute_title_roman": "III",
+        "source_url": "https://test.gov/test3.xml"
+    },
+    {
         "section": "99999",
         "title": 44,
         "usc": "abcdef",
