+++ solution/backend/regulations/tests/test_statute_links.py
@@ -167,12 +167,12 @@ class StatuteConvertersAPITestCase(APITestCase):
     def test_ssa(self):
         response = self.client.get("/v3/statutes?act=Social Security Act")
         self.assertEqual(status.HTTP_200_OK, response.status_code)
-        self.assertEqual(response.data, self.objects[1:4])
+        self.assertEqual(response.data, self.objects[1:5])
 
     def test_act_and_title(self):
         response = self.client.get("/v3/statutes?act=Social Security Act&title=3")
         self.assertEqual(status.HTTP_200_OK, response.status_code)
-        self.assertEqual(response.data, self.objects[2:3])
+        self.assertEqual(response.data, self.objects[2:4])
 
     def test_title_no_act(self):
         response = self.client.get("/v3/statutes?title=3")
