+++ solution/backend/regulations/views/statutes.py
@@ -38,7 +38,7 @@ class StatuteLinkConverterViewSet(viewsets.ReadOnlyModelViewSet):
             queryset = queryset.filter(act__iexact=act)
         if title:
             queryset = queryset.filter(statute_title__iexact=title)
-        return queryset.order_by("act", "statute_title")
+        return queryset.order_by("act", "statute_title", "usc_sort")
 
 
 class ActListSerializer(serializers.Serializer):
